Significant formatting update. Renamed at45 to nvm.

+++ drivers/spi/targets/at91sam4s/spi.c
@@ -4,54 +4,54 @@
 
 void spi_configure() {
 
-	/* ~ Enable the SPI in the Peripheral Clock Enable Register. ~ */
+	/* Enable the SPI in the Peripheral Clock Enable Register. */
 	set_bit_in_port(AT91C_ID_SPI, AT91C_BASE_PMC -> PMC_PCER);
 
-	/* ~ Select PIO A. ~ */
+	/* Select PIO A. */
 	set_bits_in_port_with_mask(AT91C_BASE_PIOA -> PIO_ASR, (AT91C_PA12_MISO | AT91C_PA13_MOSI | AT91C_PA14_SPCK));
 
-	/* ~ Initalize the SPI in master mode. ~ */
+	/* Initalize the SPI in master mode. */
 	set_bits_in_port_with_mask(AT91C_BASE_SPI -> SPI_MR, AT91C_SPI_MSTR);
 
-	/* ~ Disable communications and reset the SPI. ~ */
+	/* Disable communications and reset the SPI. */
 	set_bits_in_port_with_mask(AT91C_BASE_SPI -> SPI_CR, AT91C_SPI_SPIDIS);
 
-	/* ~ Disable the PDMAC for now. ~ */
+	/* Disable the PDMAC for now. */
 	AT91C_BASE_SPI -> SPI_PTCR = (AT91C_PDC_RXTDIS | AT91C_PDC_TXTDIS);
 
-	/* ~ Configure the SPI as a master: fixed peripheral selection: no fault detection: deselect all peripherals.  ~ */
+	/* Configure the SPI as a master: fixed peripheral selection: no fault detection: deselect all peripherals.  */
 	set_bits_in_port_with_mask(AT91C_BASE_SPI -> SPI_MR, (AT91C_SPI_PS_FIXED | (AT91C_SPI_PCS & (1 << 16)) | AT91C_SPI_MODFDIS));
 
-	/* ~ SPI_DATA_MODE_3: don't change CS after transfer: 8 bits per transfer: ~ */
+	/* SPI_DATA_MODE_3: don't change CS after transfer: 8 bits per transfer: */
 	AT91C_SPI_CSR[1] = (AT91C_SPI_CPOL | AT91C_SPI_BITS_8 | AT91C_SPI_CSAAT | (AT91C_SPI_SCBR & (128 << 8) | (AT91C_SPI_DLYBCT & (1 << 24))));
 
 }
 
 void spi_enable(void) {
 
-	/* ~ Disable use of the SPI pins by the PIO. They are now in use by the SPI controller. ~ */
+	/* Disable use of the SPI pins by the PIO. They are now in use by the SPI controller. */
 	clear_bits_in_port_with_mask(AT91C_BASE_PIOA -> PIO_PER, (AT91C_PA12_MISO | AT91C_PA13_MOSI | AT91C_PA14_SPCK));
 	set_bits_in_port_with_mask(AT91C_BASE_PIOA -> PIO_PDR, (AT91C_PA12_MISO | AT91C_PA13_MOSI | AT91C_PA14_SPCK));
 
-	/* ~ Enable the SPI. ~ */
+	/* Enable the SPI. */
 	clear_bits_in_port_with_mask(AT91C_BASE_SPI -> SPI_CR, AT91C_SPI_SPIDIS);
 	set_bits_in_port_with_mask(AT91C_BASE_SPI -> SPI_CR, AT91C_SPI_SPIEN);
 
-	/* ~ Wait for the SPI to boot. ~ */
+	/* Wait for the SPI to boot. */
 	while (!((AT91C_BASE_SPI -> SPI_SR) & AT91C_SPI_SPIENS));
 
 }
 
 void spi_disable(void) {
 
-	/* ~ Wait until the SPI has finished transmitting any data. ~ */
+	/* Wait until the SPI has finished transmitting any data. */
 	while (!((AT91C_BASE_SPI -> SPI_SR) & AT91C_SPI_TXEMPTY));
 
-	/* ~ Disable the SPI. ~ */
+	/* Disable the SPI. */
 	clear_bits_in_port_with_mask(AT91C_BASE_SPI -> SPI_CR, AT91C_SPI_SPIEN);
 	set_bits_in_port_with_mask(AT91C_BASE_SPI -> SPI_CR, AT91C_SPI_SPIDIS);
 
-	/* ~ Disable use of the SPI pins by the SPI. They are now in use by the PIO controller. ~ */
+	/* Disable use of the SPI pins by the SPI. They are now in use by the PIO controller. */
 	clear_bits_in_port_with_mask(AT91C_BASE_PIOA -> PIO_PDR, (AT91C_PA12_MISO | AT91C_PA13_MOSI | AT91C_PA14_SPCK));
 	set_bits_in_port_with_mask(AT91C_BASE_PIOA -> PIO_PER, (AT91C_PA12_MISO | AT91C_PA13_MOSI | AT91C_PA14_SPCK));
 
@@ -87,16 +87,16 @@ uint8_t spi_get(void) {
 
 }
 
-void spi_push(void *source, uint32_t length) {
+void spi_push(void *source, size_t length) {
 
-	/* ~ THIS SHOULD BE OPTIMIZED TO USE THE DMAC. ~ */
+	/* THIS SHOULD BE OPTIMIZED TO USE THE DMAC. */
 	while (length --) spi_put(*(uint8_t *)(source ++));
 
 }
 
-void spi_pull(void *destination, uint32_t length) {
+void spi_pull(void *destination, size_t length) {
 
-	/* ~ THIS SHOULD BE OPTIMIZED TO USE THE DMAC. ~ */
+	/* THIS SHOULD BE OPTIMIZED TO USE THE DMAC. */
 	while (length --) *(uint8_t *)(destination ++) = spi_get();
 
 }
