+++ drivers/usart/targets/at91sam4s/usart.c
@@ -4,16 +4,16 @@
 
 void usart_configure(AT91S_USART *usart, uint16_t baud) {
 
-	/* ~ Reset the USART controller and temporarily disable communications. ~ */
+	/* Reset the USART controller and temporarily disable communications. */
 	set_bits_in_port_with_mask(usart -> US_CR, (AT91C_US_RSTRX | AT91C_US_RSTTX | AT91C_US_RXDIS | AT91C_US_TXDIS));
 
-	/* ~ Configure the USART controller. (8N1) ~ */
+	/* Configure the USART controller. (8N1) */
 	set_bits_in_port_with_mask(usart -> US_MR, (AT91C_US_USMODE_NORMAL | AT91C_US_CLKS_CLOCK | AT91C_US_CHRL_8_BITS | AT91C_US_PAR_NONE | AT91C_US_NBSTOP_1_BIT));
 
-	/* ~ Set the baudrate. ~ */
+	/* Set the baudrate. */
 	usart -> US_BRGR = baud;
 
-	/* ~ Enable communications. ~ */
+	/* Enable communications. */
 	set_bits_in_port_with_mask(usart -> US_CR, (AT91C_US_RXEN | AT91C_US_TXEN));
 
 }
@@ -38,10 +38,10 @@ uint8_t usart_ready(AT91S_USART *usart) {
 
 void usart_put_byte(AT91S_USART *usart, uint8_t byte) {
 
-	/* ~ Wait until the USART is ready to transmit. ~ */
+	/* Wait until the USART is ready to transmit. */
 	while (!(usart -> US_CSR & AT91C_US_TXRDY));
 
-	/* ~ Send the byte. ~ */
+	/* Send the byte. */
 	usart -> US_THR = byte;
 
 }
@@ -50,36 +50,36 @@ uint8_t usart_get_byte(AT91S_USART *usart) {
 
 	while (!(usart -> US_CSR & AT91C_US_RXRDY));
 
-	/* ~ Read the byte. ~ */
+	/* Read the byte. */
 	return (usart -> US_RHR);
 
 }
 
 void usart_push(AT91S_USART *usart, void *source, size_t length) {
 
-	/* ~ THIS SHOULD BE OPTIMIZED TO USE THE DMAC. ~ */
+	/* THIS SHOULD BE OPTIMIZED TO USE THE DMAC. */
 	while (length --) usart_put_byte(usart, *(uint8_t *)(source ++));
 
 }
 
 void usart_pull(AT91S_USART *usart, void *destination, size_t length) {
 
-	/* ~ THIS SHOULD BE OPTIMIZED TO USE THE DMAC. ~ */
+	/* THIS SHOULD BE OPTIMIZED TO USE THE DMAC. */
 	while (length --) *(uint8_t *)(destination ++) = usart_get_byte(usart);
 
 }
 
-/* ~ ----------------------- USART0 ----------------------- ~ */
+/* ----------------------- USART0 ----------------------- */
 
 void usart0_configure(void *baud) {
 
-	/* ~ Enable USART0 in the PCER (peripheral clock enable register). ~ */
+	/* Enable USART0 in the PCER (peripheral clock enable register). */
 	set_bit_in_port(AT91C_ID_US0, AT91C_BASE_PMC -> PMC_PCER);
 
-	/* ~ Disable use of the RX and TX pins by the peripheral IO controller. They are now in use by the USART controller. ~ */
+	/* Disable use of the RX and TX pins by the peripheral IO controller. They are now in use by the USART controller. */
 	set_bits_in_port_with_mask(AT91C_BASE_PIOA -> PIO_PDR, (AT91C_PA5_RXD0 | AT91C_PA6_TXD0));
 
-	/* ~ Configure the USART hardware. ~ */
+	/* Configure the USART hardware. */
 	usart_configure(AT91C_BASE_US0, (uint32_t)(baud));
 
 }
@@ -114,29 +114,29 @@ uint8_t usart0_get(void) {
 
 }
 
-void usart0_push(void *source, uint32_t length) {
+void usart0_push(void *source, size_t length) {
 
 	usart_push(AT91C_BASE_US0, source, length);
 
 }
 
-void usart0_pull(void *destination, uint32_t length) {
+void usart0_pull(void *destination, size_t length) {
 
 	usart_pull(AT91C_BASE_US0, destination, length);
 
 }
 
-/* ~ ----------------------- USART1 ----------------------- ~ */
+/* ----------------------- USART1 ----------------------- */
 
 void usart1_configure(void *baud) {
 
-	/* ~ Enable USART1 in the PCER (peripheral clock enable register). ~ */
+	/* Enable USART1 in the PCER (peripheral clock enable register). */
 	set_bit_in_port(AT91C_ID_US1, AT91C_BASE_PMC -> PMC_PCER);
 
-	/* ~ Disable use of the RX and TX pins by the peripheral IO controller. They are now in use by the USART controller. ~ */
+	/* Disable use of the RX and TX pins by the peripheral IO controller. They are now in use by the USART controller. */
 	set_bits_in_port_with_mask(AT91C_BASE_PIOA -> PIO_PDR, (AT91C_PA21_RXD1 | AT91C_PA22_TXD1));
 
-	/* ~ Configure the USART hardware. ~ */
+	/* Configure the USART hardware. */
 	usart_configure(AT91C_BASE_US1, (uint32_t)(baud));
 
 }
@@ -171,19 +171,19 @@ uint8_t usart1_get(void) {
 
 }
 
-void usart1_push(void *source, uint32_t length) {
+void usart1_push(void *source, size_t length) {
 
 	usart_push(AT91C_BASE_US1, source, length);
 
 }
 
-void usart1_pull(void *destination, uint32_t length) {
+void usart1_pull(void *destination, size_t length) {
 
 	usart_pull(AT91C_BASE_US1, destination, length);
 
 }
 
-/* ~ ----------------------- DBGU ----------------------- ~ */
+/* ----------------------- DBGU ----------------------- */
 
 void dbgu_configure(void *baud) {
 
@@ -213,10 +213,10 @@ uint8_t dbgu_get(void) {
 
 }
 
-void dbgu_push(void *source, uint32_t length) {
+void dbgu_push(void *source, size_t length) {
 
 }
 
-void dbgu_pull(void *destination, uint32_t length) {
+void dbgu_pull(void *destination, size_t length) {
 
 }
